requirements.txt
adafruit-circuitpython-neopixel>=6.3.0
rpi-ws281x>=5.0.0
adafruit-blinka>=8.0.0
